requirements.txt
jupyter>=1.1.1
pandas>=2.3.3
spacy>=3.8.11
numpy
transformers
python-dotenv>=1.0.0
lyricsgenius>=3.0.0